Parse top-level import during mkpkg
